jruby: rename lib/jruby/math.rb to operators.rb
