+++ lib/nmatrix/jruby/operators.rb
